+++ pmd-kotlin/src/main/java/net/sourceforge/pmd/lang/kotlin/ast/KotlinTerminalNode.java
@@ -13,11 +13,9 @@ import net.sourceforge.pmd.lang.rule.xpath.NoAttribute;
 
 public final class KotlinTerminalNode extends BaseAntlrTerminalNode<KotlinNode> implements KotlinNode {
 
-
     KotlinTerminalNode(Token token) {
         super(token);
     }
-
 
     @Override
     public @NonNull String getText() {
@@ -48,7 +46,6 @@ public final class KotlinTerminalNode extends BaseAntlrTerminalNode<KotlinNode> 
     public String getXPathNodeName() {
         return KotlinParser.DICO.getXPathNameOfToken(getFirstAntlrToken().getType());
     }
-
 
     @Override
     public <P, R> R acceptVisitor(AstVisitor<? super P, ? extends R> visitor, P data) {