+++ pmd-kotlin/src/main/java/net/sourceforge/pmd/lang/kotlin/types/InternalApiBridge.java
@@ -25,14 +25,14 @@ public final class InternalApiBridge {
 
     private InternalApiBridge() {}
 
-    /** @see KotlinNodeTypeData#setTypeName(KotlinNode, String) */
-    public static void setTypeName(KotlinNode node, String typeName) {
-        KotlinNodeTypeData.setTypeName(node, typeName);
+    /** @see KotlinNodeTypeData#setType(KotlinNode, KotlinTypeName) */
+    public static void setType(KotlinNode node, KotlinTypeName type) {
+        KotlinNodeTypeData.setType(node, type);
     }
 
-    /** @see KotlinNodeTypeData#setReturnTypeName(KotlinNode, String) */
-    public static void setReturnTypeName(KotlinNode node, String returnTypeName) {
-        KotlinNodeTypeData.setReturnTypeName(node, returnTypeName);
+    /** @see KotlinNodeTypeData#setReturnType(KotlinNode, KotlinTypeName) */
+    public static void setReturnType(KotlinNode node, KotlinTypeName returnType) {
+        KotlinNodeTypeData.setReturnType(node, returnType);
     }
 
     public static void setAnnotationFqNames(KotlinNode node, List<String> fqnList) {