+++ pmd-kotlin/src/test/resources/net/sourceforge/pmd/lang/kotlin/rule/xpath/matchesSig/SystemPrintlnUsage.kt
@@ -1,0 +1,10 @@
+class SystemPrintlnUsage {
+    fun bad() {
+        System.out.println("debug")
+        System.err.println("error")
+    }
+
+    fun good(value: Int) {
+        val result = value * 2
+    }
+}