+++ pmd-kotlin/src/test/resources/net/sourceforge/pmd/lang/kotlin/rule/xpath/typeis/StringBufferFieldUsage.kt
@@ -1,0 +1,8 @@
+class StringBufferFieldUsage {
+    private val sb: StringBuffer = StringBuffer()
+    private val builder: StringBuilder = StringBuilder()
+
+    fun good() {
+        val local: StringBuilder = StringBuilder()
+    }
+}